# docs/swagger.yaml
definitions:
  views.AddPlanetRequest:
    properties:
      clockwise:
        type: boolean
      initial_degrees:
        type: number
      name:
        type: string
      radio:
        type: number
      solar_system_id:
        type: integer
      speed_by_day:
        type: number
    type: object
  views.AddSolarSystemRequest:
    properties:
      name:
        type: string
    type: object
  views.BaseResponse:
    properties:
      data:
        type: object
      status_code:
        type: integer
    type: object
info:
  contact: {}
  license: {}
paths:
  /planets:
    get:
      description: Gets all added planets
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/views.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/views.BaseResponse'
    post:
      description: Adds a planet into the database
      parameters:
      - description: Add Planet Request
        in: body
        name: message
        schema:
          $ref: '#/definitions/views.AddPlanetRequest'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/views.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/views.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/views.BaseResponse'
  /planets/forecast:
    get:
      description: Obtains a previously generated forecast
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/views.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/views.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/views.BaseResponse'
  /solar_systems:
    get:
      description: Gets all added solar systems
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/views.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/views.BaseResponse'
    post:
      description: Adds a solar system into the database
      parameters:
      - description: Add Solar System Request
        in: body
        name: message
        schema:
          $ref: '#/definitions/views.AddSolarSystemRequest'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/views.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/views.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/views.BaseResponse'
  /solar_systems/:
    post:
      description: Generates forecasts given added system and planets
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/views.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/views.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/views.BaseResponse'
swagger: "2.0"